backend/requirements.txt
flask
supabase
python-dotenv
PyJWT